+++ raycast-env.d.ts
@@ -17,18 +17,20 @@ declare namespace Preferences {
   export type Index = ExtensionPreferences & {
   /** API Key - You can find your API key on the remove.bg Dashboard. */
   "apiKey": string,
+  /** Default Save Location - Default location to save processed images */
+  "defaultSaveLocation": "same" | "desktop" | "downloads",
   /** Crop Image to Content - Automatically crop image to content boundaries */
   "cropToContent": boolean,
-  /** Filename Format - How to modify output filenames */
-  "filenameFormat": "nochange" | "prepend" | "append",
-  /** Filename Modifier - Text to add to filenames */
-  "fileNameModifier": string,
-  /** Default Size - Default output size */
-  "defaultSize": "preview" | "auto" | "full" | "50MP",
   /** Default Object Type - Default object type to extract */
   "defaultType": "auto" | "person" | "product" | "car" | "animal" | "graphic" | "transportation",
+  /** Default Size - Default output size */
+  "defaultSize": "preview" | "auto" | "full" | "50MP",
   /** Default Format - Default output format */
-  "defaultFormat": "png" | "jpg" | "zip"
+  "defaultFormat": "png" | "jpg" | "zip",
+  /** Modify Filename - How to modify output filenames */
+  "filenameFormat": "nochange" | "prepend" | "append",
+  /** Modifying Text - Text to add to filenames */
+  "filenameModifier": string
 }
 }